\name{MaxPro-package}
\alias{MaxPro-package}
\docType{package}
\title{
MaxPro package
}
\description{
Generate a maximum projection (MaxPro) design, a MaxPro Latin hypercube design or improve an initial design based on the MaxPro criterion
}
\details{
\tabular{ll}{
Package: \tab MaxPro\cr
Type: \tab Package\cr
Version: \tab 3.1-2\cr
Date: \tab 2015-01-27\cr
License: \tab LGPL-2.1\cr
}
Important functions in this package are: \code{\link{MaxProLHD}} generates the MaxPro Latin hypercube design, \code{\link{MaxProImproveLHD}} improves a given Latin hypercube design based on the MaxPro criterion and \code{\link{MaxPro}} searches for the locally optimal MaxPro design in the neighborhood of a given initial design. 
}
\author{
Shan Ba and V. Roshan Joseph

Maintainer: Shan Ba <shanbatr@gmail.com>
}
\references{
Joseph, V. R., Gul, E., and Ba, S. (2015) "Maximum Projection Designs for Computer Experiments," \emph{Biometrika}.
}

\keyword{ package }

